first pass to resolve circular deps
